Requested Updates to PR

+++ packages/expo-camera/android/src/main/java/expo/modules/camera/ExpoCameraView.java
@@ -233,7 +233,7 @@ public class ExpoCameraView extends CameraView implements LifecycleEventListener
       }
 
       if (options.get(VIDEO_BITRATE_KEY) != null) {
-        profile.videoBitRate = (int) options.get(VIDEO_BITRATE_KEY);
+        profile.videoBitRate = ((Double) options.get(VIDEO_BITRATE_KEY)).intValue();
       }
 
       Boolean muteValue = (Boolean) options.get(MUTE_KEY);
